+++ src/web/src/app/components/sidebar-nav/sidebar-nav.component.ts
@@ -1,14 +1,14 @@
-import { Component, OnInit, AfterViewInit, Input, ElementRef, ViewChildren, QueryList} from '@angular/core';
+import { Component, OnInit, AfterViewInit, Input, ElementRef} from '@angular/core';
 import { Router, ActivatedRoute, NavigationEnd } from '@angular/router';
 import { PagesService } from '../../shared/pages.service';
+import { HelpersService } from '../../shared/helpers.service';
 
 @Component({
   selector: 'sidebar-nav',
   templateUrl: './sidebar-nav.component.html',
-  providers: [PagesService]
+  providers: [PagesService, HelpersService]
 })
 export class SidebarNavComponent implements OnInit, AfterViewInit {
-  @ViewChildren('expandableList') expandableList: QueryList<any>;
   @Input() sidebarData;
   @Input() section;
   public sidebarNav: any;
@@ -16,13 +16,14 @@ export class SidebarNavComponent implements OnInit, AfterViewInit {
   public level_2: boolean;
   public loading: boolean;
   public expandedLevel1: any = [];
-  public expandedLevel2: any = [];
+  public halp: any = [];
 
   constructor(
     private pagesService: PagesService,
     private router: Router,
     private route: ActivatedRoute,
-    private elRef: ElementRef
+    private elRef: ElementRef,
+    private helpers: HelpersService
   ) {}
 
   ngOnInit() {}
@@ -61,12 +62,13 @@ export class SidebarNavComponent implements OnInit, AfterViewInit {
                   this.sectionTitle = item.title;
                   return thisChild.menu_order > nextChild.menu_order ? 1 : -1;
                 });
-                this.closeAccordionsMobile(this.sidebarNav);
+                this.helpers.closeAccordionsMobile(this.sidebarNav, this.expandedLevel1);
               }
             });
           },
-          () => {
+          (err) => {
             this.loading = false;
+            console.error(err);
           },
           () => {
             this.loading = false;
@@ -79,21 +81,6 @@ export class SidebarNavComponent implements OnInit, AfterViewInit {
 
   private toggleAccordion(i) {
     this.expandedLevel1[i] = !this.expandedLevel1[i];
-  }
-
-  private closeAccordionsMobile(sidebarNav) {
-
-    const checkViewport = (vp) => {
-      if (!vp.matches) {
-          setTimeout(() => {
-            this.expandedLevel1 = sidebarNav.map(i => true);
-          });
-      };
-    };
-
-    let viewport = window.matchMedia('(min-width: 600px)');
-    checkViewport(viewport);
-
   }
 
 }
